refactor[#501] : 바텀 시트 파일 네이밍 변경
