+++ src/widgets/map/ui/index.ts
@@ -4,7 +4,7 @@ export * from "./mapControlWidget";
 export * from "./mapMarkerWidget";
 export * from "./mapInitializer";
 export * from "./mapBottomSheet";
-export * from "./localMarkingList";
+export * from "./localMarkingListBottomSheet";
 export * from "./placeMarkingListBottomSheet";
-export * from "./myMarkingListBottmSheet";
+export * from "./myMarkingListBottomSheet";
 export * from "./myActivityList";
